Rename WebspaceStore to webspaceStore
