+++ src/Sulu/Bundle/PageBundle/Resources/js/containers/TeaserSelection/index.js
@@ -1,6 +1,6 @@
 // @flow
 import TeaserSelection from './TeaserSelection';
-import teaserProviderRegistry from './registries/TeaserProviderRegistry';
+import teaserProviderRegistry from './registries/teaserProviderRegistry';
 
 export default TeaserSelection;
 export {teaserProviderRegistry};
