+++ src/Sulu/Bundle/PageBundle/Resources/js/containers/TeaserSelection/registries/teaserProviderRegistry.js
