+++ src/Sulu/Bundle/PageBundle/Resources/js/containers/TeaserSelection/tests/registries/TeaserProviderRegistry.test.js
@@ -1,5 +1,5 @@
 // @flow
-import teaserProviderRegistry from '../../registries/TeaserProviderRegistry';
+import teaserProviderRegistry from '../../registries/teaserProviderRegistry';
 
 const defaultTeaserProviderOptions = {
     displayProperties: [],
