--- src/Sulu/Bundle/PageBundle/Resources/js/stores/WebspaceStore/index.js
@@ -1,4 +1,0 @@
-// @flow
-import WebspaceStore from './WebspaceStore';
-
-export default WebspaceStore;
