+++ src/Sulu/Bundle/PageBundle/Resources/js/stores/webspaceStore/index.js
@@ -1,0 +1,4 @@
+// @flow
+import WebspaceStore from './webspaceStore';
+
+export default WebspaceStore;
