+++ src/Sulu/Bundle/PageBundle/Resources/js/stores/webspaceStore/types.js
