+++ src/Sulu/Bundle/PageBundle/Resources/js/stores/webspaceStore/webspaceStore.js
