+++ src/Sulu/Bundle/PageBundle/Resources/js/types.js
@@ -1,4 +1,4 @@
 // @flow
-import type {Webspace} from './stores/WebspaceStore/types';
+import type {Webspace} from './stores/webspaceStore/types';
 
 export type {Webspace};
